Keep css extension
By keeping the extension, the css plugin can be added to the stubModules r.js definition.

The current solution strips the .css extension, generating the following output:

```js
define('css!Application/assets/css/app',[],function(){});
```

Then if a module requires this dep as follows:

```js
    'css!./assets/css/app.css'
```

require/almond will request the css plugin to fetch the module.
This works fine, but the css plugin must be in the source for this to work. This means it can't be stubbed.

This fixes the issue because, by keeping the extension, require/almond know that the module is already defined and will not hit the css plugin.